requirements.txt
asgiref==3.9.1
Django==5.2.4
iniconfig==2.1.0
packaging==25.0
pillow==11.3.0
pluggy==1.6.0
Pygments==2.19.2
pytest==8.4.1
pytest-django==4.11.1
sqlparse==0.5.3
